test: move paths tests into dedicated file
